Splits Yaml protocols in different traits by domain (sut, model section); updates unit tests accordingly.
